Make gaps in threads scale with line width
Subtracting 1 is not a good way to produce gaps between threads, because
as we add more threads, the width of each thread shrinks to fit in the
canvas, but the space between them remains a fixed number of pixels.
This causes a sett to grow darker (as there's more gap space relative to
color space) as we increase the number of threads in the pattern, which
makes the colors appear less vibrant.  Eventually, we get threads so
narrow that the gaps are wider than the threads, and negative thread
width makes no sense.

Instead, let's use fractions -- using 80% of the space for each row or
column leaves a 10% gap on either side of the thread, which looks pretty
okay.